+++ src/Formatter/CrossReferencesConfigurator.php
@@ -101,15 +101,15 @@ class CrossReferencesConfigurator
             <xsl:choose>
                 <xsl:when test="@unknown = 1">
                     <span class="DiscussionLink DiscussionUnknown">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if>
                     </span>
                 </xsl:when>
                 <xsl:otherwise>
                     <a href="{$DISCUSSION_URL}{@id}" class="DiscussionLink">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if>
                     </a>
                 </xsl:otherwise>
@@ -134,15 +134,15 @@ class CrossReferencesConfigurator
             <xsl:choose>
                 <xsl:when test="@unknown = 1">
                     <span class="DiscussionLink DiscussionUnknown">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if>
                     </span>
                 </xsl:when>
                 <xsl:otherwise>
                     <a href="{@url}" class="DiscussionLink">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if>
                     </a>
                 </xsl:otherwise>
@@ -168,15 +168,15 @@ class CrossReferencesConfigurator
             <xsl:choose>
                 <xsl:when test="@unknown = 1">
                     <span class="DiscussionLink DiscussionUnknown">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if> <span class="DiscussionComment">(<xsl:value-of select="@comment"/>)</span>
                     </span>
                 </xsl:when>
                 <xsl:otherwise>
                     <a href="{@url}" class="DiscussionLink">
-                        <xsl:value-of select="@title"/> <xsl:if test="$SHOW_DISCUSSION_ID = 1">
-                            <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
+                        <xsl:value-of select="@title"/>
+                        <xsl:if test="$SHOW_DISCUSSION_ID = 1"> <span class="DiscussionId">#<xsl:value-of select="@id"/></span>
                         </xsl:if> <span class="DiscussionComment">(<xsl:value-of select="@comment"/>)</span>
                     </a>
                 </xsl:otherwise>