+++ tests/unit/CrossReferencesConfiguratorTest.php
@@ -24,6 +24,8 @@
 namespace Club1\CrossReferences\Tests\unit;
 
 use Club1\CrossReferences\Formatter\CrossReferencesConfigurator;
+use DOMDocument;
+use DOMElement;
 use Flarum\Foundation\ErrorHandling\LogReporter;
 use Flarum\Http\RouteCollection;
 use Flarum\Http\UrlGenerator;
@@ -35,9 +37,13 @@ use Mockery as m;
 use Mockery\MockInterface;
 use s9e\TextFormatter\Configurator;
 use s9e\TextFormatter\Parser;
+use s9e\TextFormatter\Renderer;
 use s9e\TextFormatter\Utils;
 
+use function PHPUnit\Framework\assertContains;
+use function PHPUnit\Framework\assertCount;
 use function PHPUnit\Framework\assertEquals;
+use function PHPUnit\Framework\assertGreaterThanOrEqual;
 
 class CrossReferencesConfiguratorTest extends TestCase
 {
@@ -56,8 +62,8 @@ class CrossReferencesConfiguratorTest extends TestCase
     {
         parent::setUp();
         $this->settingsRepo = m::mock(SettingsRepositoryInterface::class);
-        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.show_discussion_id')->andReturn(false);
-        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.enable_short_references')->andReturn(true);
+        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.show_discussion_id')->andReturn(false)->byDefault();
+        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.enable_short_references')->andReturn(true)->byDefault();
 
         $routeCollection = m::mock(RouteCollection::class);
         $routeCollection->shouldReceive('route')->with('discussion', ['id' => ''])->andReturn('https://forum.club1.fr/d/');
@@ -96,7 +102,7 @@ class CrossReferencesConfiguratorTest extends TestCase
         $this->discussionModel->shouldReceive('find')->andReturn($discussion);
     }
 
-    public function basicConfiguration(): Parser
+    public function basicParserConfiguration(): Parser
     {
         $this->mockDiscussion();
         $xrefConfigurator = new CrossReferencesConfigurator($this->settingsRepo, $this->urlGenerator);
@@ -106,6 +112,14 @@ class CrossReferencesConfiguratorTest extends TestCase
         return $parser;
     }
 
+    public function basicRendererConfiguration(): Renderer
+    {
+        $xrefConfigurator = new CrossReferencesConfigurator($this->settingsRepo, $this->urlGenerator);
+        $xrefConfigurator($this->configurator);
+        extract($this->configurator->finalize());
+        return $renderer;
+    }
+
     /**
      * @dataProvider shortReferencesProvider
      * @param string $text The text to parse.
@@ -113,7 +127,7 @@ class CrossReferencesConfiguratorTest extends TestCase
      */
     public function testShortReferences(string $text, array $expected): void
     {
-        $parser = $this->basicConfiguration();
+        $parser = $this->basicParserConfiguration();
         $xml = $parser->parse($text);
         $actual = Utils::getAttributeValues($xml, 'CROSSREFERENCESHORT', 'id');
         assertEquals($expected, $actual);
@@ -141,7 +155,7 @@ class CrossReferencesConfiguratorTest extends TestCase
      */
     public function testUrlReferences(string $text, array $expected): void
     {
-        $parser = $this->basicConfiguration();
+        $parser = $this->basicParserConfiguration();
         $xml = $parser->parse($text);
         $actual = Utils::getAttributeValues($xml, 'CROSSREFERENCEURL', 'id');
         assertEquals($expected, $actual);
@@ -179,7 +193,7 @@ class CrossReferencesConfiguratorTest extends TestCase
      */
     public function testCommentUrlReferences(string $text, array $expected): void
     {
-        $parser = $this->basicConfiguration();
+        $parser = $this->basicParserConfiguration();
         $xml = $parser->parse($text);
         $actual = Utils::getAttributeValues($xml, 'CROSSREFERENCEURLCOMMENT', 'id');
         assertEquals($expected, $actual);
@@ -279,6 +293,88 @@ class CrossReferencesConfiguratorTest extends TestCase
         return [
             ['#42', true, ['42']],
             ['#42', false, []],
+        ];
+    }
+
+    /**
+     * @dataProvider renderingUnknownProvider
+     * @param string $xml The xml to render.
+     * @param bool $showId If the show ID parameter should be enabled.
+     * @param string $title The expected title.
+     */
+    public function testRenderingUnknown(string $xml, bool $showId, string $title): void
+    {
+        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.show_discussion_id')->andReturn($showId);
+        $renderer = $this->basicRendererConfiguration();
+        $html = $renderer->render($xml);
+        $dom = new DOMDocument();
+        $dom->loadHTML($html);
+        assertCount(0, $dom->getElementsByTagName('a'));
+        assertGreaterThanOrEqual(1, $dom->getElementsByTagName('span')->length);
+        $span = $dom->getElementsByTagName('span')->item(0);
+        assert($span instanceof DOMElement);
+        $classes = explode(' ', $span->getAttribute('class'));
+        assertContains('DiscussionLink', $classes);
+        assertContains('DiscussionUnknown', $classes);
+        assertEquals($title, $span->textContent);
+    }
+
+    public function renderingUnknownProvider(): array
+    {
+        return [
+            ['<CROSSREFERENCESHORT id="42" title="[unknown]" unknown="1">#42</CROSSREFERENCESHORT>',
+             false, '[unknown]'],
+            ['<CROSSREFERENCEURL id="42" title="[unknown]" url="https://forum.club1.fr/d/42-coucou" unknown="1"></CROSSREFERENCEURL>',
+             false, '[unknown]'],
+            ['<CROSSREFERENCEURLCOMMENT id="42" title="[unknown]" url="https://forum.club1.fr/d/42-coucou/12" unknown="1" comment="comment"></CROSSREFERENCEURLCOMMENT>',
+             false, '[unknown] (comment)'],
+            ['<CROSSREFERENCESHORT id="42" title="[unknown]" unknown="1">#42</CROSSREFERENCESHORT>',
+             true, '[unknown] #42'],
+            ['<CROSSREFERENCEURL id="2" title="[unknown]" url="https://forum.club1.fr/d/42-coucou" unknown="1"></CROSSREFERENCEURL>',
+             true, '[unknown] #2'],
+            ['<CROSSREFERENCEURLCOMMENT id="3" title="[unknown]" url="https://forum.club1.fr/d/42-coucou/12" unknown="1" comment="comment"></CROSSREFERENCEURLCOMMENT>',
+             true, '[unknown] #3 (comment)'],
+        ];
+    }
+
+    /**
+     * @dataProvider renderingValidProvider
+     * @param string $xml The xml to render.
+     * @param bool $showId If the show ID parameter should be enabled.
+     * @param string $title The expected title.
+     * @param string $url The expected url.
+     */
+    public function testRenderingValid(string $xml, bool $showId, string $title, string $url): void
+    {
+        $this->settingsRepo->shouldReceive('get')->with('club-1-cross-references.show_discussion_id')->andReturn($showId);
+        $renderer = $this->basicRendererConfiguration();
+        $html = $renderer->render($xml);
+        $dom = new DOMDocument();
+        $dom->loadHTML($html);
+        assertCount(1, $dom->getElementsByTagName('a'));
+        $a = $dom->getElementsByTagName('a')->item(0);
+        assert($a instanceof DOMElement);
+        $classes = explode(' ', $a->getAttribute('class'));
+        assertContains('DiscussionLink', $classes);
+        assertEquals($title, $a->textContent);
+        assertEquals($url, $a->getAttribute('href'));
+    }
+
+    public function renderingValidProvider(): array
+    {
+        return [
+            ['<CROSSREFERENCESHORT id="42" title="Coucou 42">#42</CROSSREFERENCESHORT>',
+             false, 'Coucou 42', 'https://forum.club1.fr/d/42'],
+            ['<CROSSREFERENCEURL id="42" title="Dummy" url="https://forum.club1.fr/d/42-coucou"></CROSSREFERENCEURL>',
+             false, 'Dummy', 'https://forum.club1.fr/d/42-coucou'],
+            ['<CROSSREFERENCEURLCOMMENT id="42" title="Coucou" url="https://forum.club1.fr/d/42-coucou/12" comment="comment"></CROSSREFERENCEURLCOMMENT>',
+             false, 'Coucou (comment)', 'https://forum.club1.fr/d/42-coucou/12'],
+            ['<CROSSREFERENCESHORT id="42" title="Coucou 42">#42</CROSSREFERENCESHORT>',
+             true, 'Coucou 42 #42', 'https://forum.club1.fr/d/42'],
+            ['<CROSSREFERENCEURL id="77" title="Dummy" url="https://forum.club1.fr/d/7-coucou"></CROSSREFERENCEURL>',
+             true, 'Dummy #77', 'https://forum.club1.fr/d/7-coucou'],
+            ['<CROSSREFERENCEURLCOMMENT id="1" title="Coucou" url="https://forum.club1.fr/d/1/12" comment="comment"></CROSSREFERENCEURLCOMMENT>',
+             true, 'Coucou #1 (comment)', 'https://forum.club1.fr/d/1/12'],
         ];
     }
 }